migrate cond api
