+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_dict.py
@@ -97,7 +97,7 @@ class MainNetWithDict(fluid.dygraph.Layer):
             ),
         }
         # TODO(Aurelius84): The following code will be converted into:
-        # max_len = layers.cond(paddle.shape(input)[0] != max_len,
+        # max_len = paddle.static.nn.cond(paddle.shape(input)[0] != max_len,
         #                       lambda: paddle.shape(input)[0], lambda: max_len)
         # But max_len should be wrapped into tensor, which is not supported.
