+++ python/paddle/fluid/tests/unittests/test_program_code.py
@@ -46,7 +46,7 @@ class TestProgramToReadableCode(unittest.TestCase):
             x = layers.fill_constant(shape=[1], dtype='float32', value=0.1)
             y = layers.fill_constant(shape=[1], dtype='float32', value=0.23)
             pred = paddle.less_than(y, x)
-            out = layers.cond(pred, true_func, false_func)
+            out = paddle.static.nn.cond(pred, true_func, false_func)
 
     def test_program_code(self):
         self.var._to_readable_code()
