+++ python/paddle/fluid/variable_index.py
@@ -341,7 +341,7 @@ def get_value_for_bool_tensor(var, item):
         var_shape[0] = 0
         return paddle.empty(var_shape, dtype=var.dtype)
 
-    from .layers.control_flow import cond
+    from paddle.static.nn import cond
 
     return cond(
         item.any(), lambda: idx_not_empty(var, item), lambda: idx_empty(var)
@@ -874,7 +874,7 @@ def set_value_for_bool_tensor(var, item, value):
         out = scatter_nd_add(var, idx, gather_val_new)
         var[:] = out
 
-    from .layers.control_flow import cond
+    from paddle.static.nn import cond
 
     # If all the bool index is False, just do nothing
     cond(item.any(), lambda: idx_not_empty(var, item, value))
