+++ python/paddle/static/nn/__init__.py
@@ -22,18 +22,19 @@ from .common import conv2d_transpose  # noqa: F401
 from .common import conv3d_transpose  # noqa: F401
 from .common import bilinear_tensor_product  # noqa: F401
 from .common import py_func  # noqa: F401
+from .common import prelu  # noqa: F401
+from .control_flow import Assert  # noqa: F401
+from .control_flow import increment  # noqa: F401
+from .control_flow import cond  # noqa: F401
+from .loss import nce  # noqa: F401
 
 from ...tensor.creation import create_parameter  # noqa: F401
 from ...fluid.layers import batch_norm  # noqa: F401
 from ...fluid.layers import case  # noqa: F401
-from ...fluid.layers import cond  # noqa: F401
 from ...fluid.layers import conv2d  # noqa: F401
 from ...fluid.layers import crf_decoding  # noqa: F401
 from ...fluid.layers import layer_norm  # noqa: F401
 from ...fluid.layers import multi_box_head  # noqa: F401
-from .loss import nce  # noqa: F401
-from .common import prelu  # noqa: F401
-from .control_flow import Assert, increment  # noqa: F401
 from ...fluid.layers import row_conv  # noqa: F401
 from ...fluid.layers import spectral_norm  # noqa: F401
 from ...fluid.layers import switch_case  # noqa: F401
